[feat] registratePage: rebase components fo shared + rebase logic inside components from page
