+++ src/shared/input/input.hbs
@@ -1,12 +1,12 @@
 <div action="{{ action }}" id="{{ formGroupID }}" class="form__group">
     {{#if label}}
-    <label for="{{ inputID }}" class="form__label">{{ label }}</label>
+    <label for="{{ id }}" class="form__label">{{ label }}</label>
     {{/if}}
     <div class="form__input-wrapper">
         {{#if preIconSrc}}
         <img class="{{ preIconClass }}" src="{{ preIconSrc }}" alt="{{ preIconClass }}">
         {{/if}}
-        <input placeholder="{{ placeholder }}" type="{{ type }}" id="{{ inputID }}" name="{{ name }}" class="form__input" required="{{ required }}">
+        <input placeholder="{{ placeholder }}" type="{{ type }}" id="{{ id }}" name="{{ name }}" class="form__input" required="{{ required }}">
         {{#if postIconSrc}}
         <img class="watch-icon" id="{{ postIconID }}" src="{{ postIconSrc }}" alt="lock">
         {{/if}}
